+++ src/printer/html.js
@@ -4,20 +4,5 @@ module.exports = createPrinter({
     print: chunk => chunk
         .replace(/&/g, '&amp;')
         .replace(/</g, '&lt;')
-        .replace(/>/g, '&gt;'),
-
-    ranges: {
-        syntax: {
-            open: (data) => '<span class="syntax--' + data + '">',
-            close: () => '</span>'
-        },
-        spotlight: {
-            open: () => '<span class="spotlight">',
-            close: () => '</span>'
-        },
-        match: {
-            open: () => '<span class="match">',
-            close: () => '</span>'
-        }
-    }
+        .replace(/>/g, '&gt;')
 });